update smart contract data type

+++ contracts/Counter.sol
@@ -2,13 +2,13 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 contract Counter {
-  int256 public counter;
+  int public counter;
 
   constructor() {
     counter = 0;
   }
 
-  function incrementCounter() public returns(int256) {
+  function incrementCounter() public returns(int) {
     return counter++;
   }
 }
